requirements.txt
flask
neo4j
passlib
scikit-learn
google-cloud-storage
gunicorn
google-auth
appengine-python-standard>=1.0.0
itsdangerous
pandas
beautifulsoup4
cfscrape
openai==0.28.1
pdfplumber
matplotlib
plotly
python-pptx
requests
requests_oauthlib
